Split parser and integration tests
